# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Sintra AI Orchestration Platform API
  description: |
    Enterprise-grade AI orchestration platform for intelligent automation and workflow management.
    
    ## Features
    - Multi-agent AI orchestration
    - Real-time chat and collaboration
    - Knowledge base management
    - Workflow automation
    - Analytics and monitoring
    
    ## Authentication
    The API uses JWT Bearer token authentication. Include the token in the Authorization header:
    ```
    Authorization: Bearer <your-jwt-token>
    ```
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests per hour
    - AI chat endpoints: 100 requests per hour
    - Admin endpoints: 500 requests per hour
    
    ## Error Handling
    All endpoints return standardized error responses with German error messages.
    
    ## SDKs
    - JavaScript/TypeScript: `@sintra-ai/sdk`
    - Python: `sintra-ai-python`
    - Go: `github.com/sintra-ai/go-sdk`
  version: 3.0.0
  contact:
    name: Sintra AI Support
    email: support@sintra-ai.com
    url: https://sintra-ai.com/support
  license:
    name: Proprietary
    url: https://sintra-ai.com/license
  termsOfService: https://sintra-ai.com/terms
servers:
  - url: https://api.sintra-ai.com/v3
    description: Production server
  - url: https://staging-api.sintra-ai.com/v3
    description: Staging server
  - url: http://localhost:3000/api
    description: Development server
security:
  - BearerAuth: []
paths:
...
  /profile:
    get:
      tags:
        - Profile
      summary: Get user profile
      description: Returns the current user's profile information
      operationId: getUserProfile
      responses:
        '200':
          description: User profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '401':
          $ref: '#/components/responses/Unauthorized'

    put:
      tags:
        - Profile
      summary: Update user profile
      description: Updates the current user's profile information
      operationId: updateUserProfile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateUserProfileRequest'
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT Bearer token authentication
  schemas:
...
    UserProfile:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier of the user
        email:
          type: string
          format: email
          description: User's email address
        name:
          type: string
          description: User's full name
        avatar:
          type: string
          format: uri
          description: URL to user's avatar image
        role:
          type: string
          enum: [user, admin, super_admin]
          description: User's role
        preferences:
          type: object
          description: User preferences
        createdAt:
          type: string
          format: date-time
        lastLoginAt:
          type: string
          format: date-time
    UpdateUserProfileRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 100
        avatar:
          type: string
          format: uri
        preferences:
          type: object
...
    Error:
      type: object
      properties:
        error:
          type: object
          properties:
            id:
              type: string
              description: Unique error identifier
            type:
              type: string
              description: Error type
            title:
              type: string
              description: Error title in German
            message:
              type: string
              description: Error message in German
            suggestion:
              type: string
              description: Suggested solution in German
            severity:
              type: string
              enum: [LOW, MEDIUM, HIGH, CRITICAL]
              description: Error severity level
            timestamp:
              type: string
              format: date-time
            requestId:
              type: string
              description: Request identifier for tracing
            context:
              type: object
              description: Additional error context
  responses:
    BadRequest:
      description: Bad request - invalid input data
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              id: "err_1640995200_abc123"
              type: "VALIDATION_ERROR"
              title: "Validierungsfehler"
              message: "Die bereitgestellten Daten sind ungültig."
              suggestion: "Bitte überprüfen Sie Ihre Eingaben und versuchen Sie es erneut."
              severity: "MEDIUM"
              timestamp: "2025-12-14T20:59:59.000Z"
    Unauthorized:
      description: Unauthorized - authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error:
              id: "err_1640995200_def456"
              type: "AUTHENTICATION_ERROR"
              title: "Authentifizierung fehlgeschlagen"
              message: "Sie müssen angemeldet sein, um auf diese Ressource zuzugreifen."
              suggestion: "Bitte melden Sie sich an und versuchen Sie es erneut."
              severity: "HIGH"
              timestamp: "2025-12-14T20:59:59.000Z"
...